# out2/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: alberto.est.jimenez@gmail.com
    name: Alberto Jimenez
  description: Documentation for the client-accessible endpoints of the StackV orchestrator.
  title: StackV Orchestrator Developer API
  version: 0.8.2
servers:
- url: https://stackv.dev3.virnao.com/StackV-web/restapi/
...
paths:
...
  /task/assigned/{assigned}:
    delete:
      description: Removes all tasks assigned to a specific agent.
      parameters:
      - description: The agent or application used as a task delineation.
        explode: false
        in: path
        name: assigned
        required: true
        schema:
          example: janus_device_manager
          type: string
        style: simple
      responses:
        "204":
          $ref: "#/components/responses/Empty"
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete assigned tasks
      tags:
      - task
      x-accepts:
      - application/json
      x-tags:
      - tag: task
    get:
      description: Queries all tasks assigned to a specific agent.
      parameters:
      - description: The agent or application used as a task delineation.
        explode: false
        in: path
        name: assigned
        required: true
        schema:
          example: janus_device_manager
          type: string
        style: simple
      - description: Whether to retrieve archived tasks as well.
        explode: true
        in: query
        name: archived
        required: false
        schema:
          default: false
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/paths/~1task~1uuid~1%7Buuid%7D/get/responses/200/content/application~1json/schema"
                type: array
          description: List of task entities.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve all assigned tasks
      tags:
      - task
      x-accepts:
      - application/json
      x-tags:
      - tag: task
  /task/uuid/{uuid}:
    delete:
      description: Removes a task by uuid.
      parameters:
      - description: The uuid of the requested task.
        explode: false
        in: path
        name: uuid
        required: true
        schema:
          example: b65e532d-4aef-46d4-b8f2-d6f3736ffd25
          type: string
        style: simple
      responses:
        "204":
          $ref: "#/components/responses/Empty"
        "500":
          $ref: "#/components/responses/Error"
      summary: Delete a task
      tags:
      - task
      x-accepts:
      - application/json
      x-tags:
      - tag: task
    get:
      description: Queries a task by uuid.
      parameters:
      - description: The uuid of the requested task.
        explode: false
        in: path
        name: uuid
        required: true
        schema:
          example: b65e532d-4aef-46d4-b8f2-d6f3736ffd25
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_task_uuid__uuid__get_200_response"
          description: A task entity.
        "500":
          $ref: "#/components/responses/Error"
      summary: Retrieve specific task
      tags:
      - task
      x-accepts:
      - application/json
      x-tags:
      - tag: task
    put:
      description: Updates a task via full update DTO.
      parameters:
      - description: The uuid of the requested task.
        explode: false
        in: path
        name: uuid
        required: true
        schema:
          example: b65e532d-4aef-46d4-b8f2-d6f3736ffd25
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SENSE_Task_Update_DTO"
        description: A task update DTO.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1task~1uuid~1%7Buuid%7D/get/responses/200/content/application~1json/schema"
          description: A task entity.
        "500":
          $ref: "#/components/responses/Error"
      summary: Update a task
      tags:
      - task
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: task
  /task/uuid/{uuid}/{status}:
    put:
      description: Updates a task's status and message.
      parameters:
      - description: The uuid of the requested task.
        explode: false
        in: path
        name: uuid
        required: true
        schema:
          example: b65e532d-4aef-46d4-b8f2-d6f3736ffd25
          type: string
        style: simple
      - description: The new status of the requested task.
        explode: false
        in: path
        name: status
        required: true
        schema:
          description: The current status of the task.
          enum:
          - PENDING
          - ACCEPTED
          - WAITING
          - FINISHED
          - FAILED
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: The new status message.
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/paths/~1task~1uuid~1%7Buuid%7D/get/responses/200/content/application~1json/schema"
          description: A task entity.
        "500":
          $ref: "#/components/responses/Error"
      summary: Update a task status
      tags:
      - task
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: task
components:
  responses:
    Error:
      description: Internal server fault or standardized error response.
    Empty:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/inline_object"
      description: "A standard empty response, usually returned on update or creation\
        \ requests."
...
  schemas:
...
    _task_uuid__uuid__get_200_response:
      description: A task stored within the SENSE orchestrator.
      example:
        archived: true
        expires: 2025-01-08T00:04:17Z
        expired: true
        created: 2024-11-06T19:58:18Z
        statusTime: 2024-11-09T14:12:45Z
        taskClass: net.maxgigapop.mrs.addon.janus.data.JANUSTaskConfig
        assigned: janus_device_manager
        id: 0.8008281904610115
        uuid: 48ab0cfc-3f50-4c4e-8c34-463156760022
        config: "{\"command\":\"add_user\",\"target\":\"nrp-04.nrp-nautilus.io\",\"\
          principal\":\"admin\",\"context\":\"b4c48872-8266-471e-ad47-903c6d447ec6\"\
          }"
        statusMessage: "{\"callbackURL\": \"random_url\"}"
        status: "{}"
      properties:
        id:
          description: Auto-generated ID.
          type: number
        uuid:
          description: The natural UUID of the task.
          example: 48ab0cfc-3f50-4c4e-8c34-463156760022
          type: string
        taskClass:
          description: The Java class that maps the task's configuration.
          example: net.maxgigapop.mrs.addon.janus.data.JANUSTaskConfig
          type: string
        assigned:
          description: The assigned agent or application. This can be likened to a
            tag or queue/topic name.
          example: janus_device_manager
          type: string
        archived:
          description: Whether the task has been archived.
          type: boolean
        created:
          description: The timestamp of the task's creation.
          example: 2024-11-06T19:58:18Z
          type: string
        expires:
          description: The timestamp of the task's expiration.
          example: 2025-01-08T00:04:17Z
          type: string
        status:
          $ref: "#/paths/~1task~1uuid~1%7Buuid%7D~1%7Bstatus%7D/put/parameters/1/schema"
        statusTime:
          description: The timestamp of the task's last status update.
          example: 2024-11-09T14:12:45Z
          type: string
        statusMessage:
          description: An optional message or JSON string to attach to any status
            update.
          example: "{\"callbackURL\": \"random_url\"}"
          type: string
        config:
          description: The immutable configuration of the task itself.
          example: "{\"command\":\"add_user\",\"target\":\"nrp-04.nrp-nautilus.io\"\
            ,\"principal\":\"admin\",\"context\":\"b4c48872-8266-471e-ad47-903c6d447ec6\"\
            }"
          type: object
        expired:
          description: Whether the task has expired.
          type: boolean
      type: object
    SENSE_Task_Update_DTO:
      description: This schema represents an update object for SENSE Task records.
      properties:
        status:
          $ref: "#/paths/~1task~1uuid~1%7Buuid%7D~1%7Bstatus%7D/put/parameters/1/schema"
        message:
          description: The optional status message to attack to this task.
          example: "[\"{\\\"foo\\\":\\\"bar\\\",\\\"test\\\":[\\\"1\\\", \\\"2\\\"\
            , \\\"3\\\"]}\",\"$callback_url\"]"
          type: string
        archived:
          description: An optional flag to archive or unarchive the task.
          type: boolean
      title: SENSE Task Update DTO
      type: object
    inline_object:
      example:
        data: {}
      properties:
        data:
          additionalProperties: false
          description: An empty object.
          example: {}
          type: object
      required:
      - data
      - meta
      type: object
...
  securitySchemes:
    OAuth2:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/auth
          scopes:
            openid: OIDC Mode.
          tokenUrl: https://auth.dev3.virnao.com/auth/realms/StackV/protocol/openid-connect/token
      type: oauth2